feat: add OpenCommit Action workflow
Add a new workflow file `.github/workflows/opencommit.yml` to enable the OpenCommit Action. This action will be triggered on push events, excluding the branches `main`, `master`, `dev`, `development`, and `release`. The workflow runs on the latest version of Ubuntu and has a timeout of 10 minutes. It requires write permissions and consists of the following steps:

1. Setup Node.js Environment: Sets up the Node.js environment with version 16.
2. Checkout: Checks out the repository with a fetch depth of 0.
3. OpenCommit: Uses the `di-sukharev/opencommit@github-action-v1.0.4` action to perform the OpenCommit process. It requires the `GITHUB_TOKEN` secret to be set in the repository's actions secrets. Additionally, the OpenAI API key is set as the `OCO_OPENAI_API_KEY` secret. Customization options include setting the maximum number of tokens, the base path, enabling or disabling description and emoji generation, selecting the OpenAI model, and specifying the language and prompt module.

chore: remove commitlint dependency

Remove the `commitlint` dependency from the `package.json` file.